refactor: migrate header actions to teleport pattern
Removes centralized HeaderActions component and distributes actions directly into each page component using Vue's Teleport feature. Each page now manages its own header controls, improving component locality and reducing prop drilling.

Benefits include better component isolation, simplified layout logic, and more maintainable code structure. The responsive behavior for mobile/desktop views is now handled within each page context rather than through conditional props.
