+++ layers/dashboard/app/pages/dashboard/link.vue
@@ -7,6 +7,11 @@ definePageMeta({
 
 const slug = useRoute().query.slug
 const linksStore = useDashboardLinksStore()
+const analysisStore = useDashboardAnalysisStore()
+
+function handleDateChange(dateRange: [number, number]) {
+  analysisStore.updateDateRange(dateRange)
+}
 
 const link = ref<Link | null>(null)
 const id = computed(() => link.value?.id)
@@ -39,6 +44,16 @@ linksStore.onLinkUpdate(({ link: updatedLink, type }) => {
 
 <template>
   <main class="space-y-6">
+    <Teleport to="#dashboard-header-actions" defer>
+      <div
+        class="
+          flex-1
+          sm:hidden
+        "
+      />
+      <DashboardDatePicker @update:date-range="handleDateChange" />
+    </Teleport>
+
     <DashboardLinksLink
       v-if="link?.id"
       :link="link"
