+++ layers/dashboard/app/pages/dashboard/links.vue
@@ -6,6 +6,20 @@ definePageMeta({
 
 <template>
   <main class="space-y-6">
+    <Teleport to="#dashboard-header-actions" defer>
+      <LazyDashboardLinksEditor />
+      <div
+        class="
+          flex-1
+          sm:hidden
+        "
+      />
+      <DashboardLinksSort />
+      <LazyDashboardLinksSearch
+        class="max-sm:w-full"
+      />
+    </Teleport>
+
     <DashboardLinks />
   </main>
 </template>
